adding yang ketinggalan

+++ tables/table-orders/table-orders.sql
@@ -1,7 +1,7 @@
 CREATE TABLE orders(
 	id INT NOT NULL AUTO_INCREMENT,
 	amount INT NOT NULL,
-	total_price INT NOT NULL,
+	total_price DECIMAL(5,2) NOT NULL,
 	customer_id INT,
 	product_id INT,
 PRIMARY KEY(id),
